requirements.txt
# Core
pytz==2022.1
Django==4.0.4
django-configurations==2.3.2
gunicorn==20.1.0
newrelic==7.10.0.175

# For the persistence stores
psycopg2-binary==2.9.3
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.13.1
Markdown==3.3.7
django-filter==21.1
django-cors-headers==3.13.0

# Developer Tools
ipdb==0.13.9
ipython==8.3.0
mkdocs==1.3.0
flake8==4.0.1

# Testing
mock==4.0.3
factory-boy==3.2.1
django-nose==1.4.7
nose-progressive==1.5.2
coverage==6.4
model_bakery==1.6.0

# Static and Media Storage
django-storages==1.12.3
boto3==1.23.5

